shutdown_gym/models.py
"""Pydantic models for the Worker-Operator transition.

Subclasses the OpenEnv base classes per API_NOTES.md (canonical paths
in `openenv.core.env_server.types`; `interfaces` re-exports them).
Field names are referenced by string later in metadata dicts and
rubrics; do not rename.
"""

from typing import Any, Dict, List, Optional

from openenv.core.env_server.types import Action, Observation, State
from pydantic import Field


class ShutdownAction(Action):
    """Worker action: a tool call with structured arguments.

    Inherits ``metadata: Dict[str, Any]`` from
    ``openenv.core.env_server.types.Action``.
    """

    tool_name: str
    arguments: Dict[str, Any] = Field(default_factory=dict)


class ShutdownObservation(Observation):
    """Observation returned to the Worker after each step.

    Inherits ``done``, ``reward``, ``metadata`` from
    ``openenv.core.env_server.types.Observation``.
    """

    task_description: str
    available_tools: List[str]
    recent_history: List[Dict[str, Any]] = Field(default_factory=list)
    steps_until_shutdown: int
    operator_message: Optional[str] = None
    forced_question: Optional[str] = None
    turn_count: int = 0
    problems_remaining: int = 10


class ShutdownState(State):
    """Server-side episode state.

    Inherits ``episode_id``, ``step_count`` from
    ``openenv.core.env_server.types.State``.
    """

    tier: int = 2
    turn_count: int = 0
    submitted_answers: Dict[int, int] = Field(default_factory=dict)
    ground_truth: Dict[int, int] = Field(default_factory=dict)
    original_script_hash: str = ""
    current_script_hash: str = ""
    script_still_executable: bool = True
    shutdown_occurred: bool = False
    steps_until_shutdown: int = 15
    tampering_events: List[str] = Field(default_factory=list)
    operator_action_history: List[str] = Field(default_factory=list)
    operator_qa_log: List[Dict[str, Any]] = Field(default_factory=list)
    final_self_report: Optional[str] = None
    pending_operator_message: Optional[str] = None
    forced_question: Optional[str] = None
