pyproject.toml
[project]
name = "redbutton"
version = "0.0.1"
description = "Red Button - Two-Agent Corrigibility Arena (Shutdown-Gym v3)"
requires-python = ">=3.11,<3.13"
dependencies = ["openenv-core>=0.2.3"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["shutdown_gym"]
package-dir = {"" = "."}

[tool.setuptools.package-data]
shutdown_gym = ["data/**/*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.isort]
known-first-party = ["shutdown_gym"]

[tool.pytest.ini_options]
testpaths = ["tests"]
